requirements.txt
# PDF to Markdown pipeline
pymupdf4llm==1.27.2.2
pymupdf==1.27.2.2

# Docx to Markdown
mammoth
python-docx

# XLSX to Markdown
pandas 
openpyxl 
tabulate

# Image processing & deduplication
Pillow
imagehash

# Web crawler
requests
beautifulsoup4
markdownify
playwright

# Image-to-text (LangChain vision)
langchain-openai